chore: refactor DialogService
